Fruit: use floats for position coordinates

+++ include/Fruit.hpp
@@ -13,21 +13,21 @@ public:
      * @param x The fruit's position along the x-axis
      * @param y The fruit's position along the y-axis
      */
-    Fruit(int x, int y) noexcept;
+    Fruit(float x, float y) noexcept;
 
     /**
      * @brief Get the position of the fruit
      * 
      * @return sf::Vector2i const& The position of the fruit as a const lvalue reference
      */
-    sf::Vector2i const& getPosition() const& noexcept;
+    sf::Vector2f const& getPosition() const& noexcept;
     
     /**
      * @brief Get the position of the fruit
      * 
      * @return sf::Vector2i The position of the fruit as an rvalue
      */
-    sf::Vector2i getPosition() && noexcept;
+    sf::Vector2f getPosition() && noexcept;
 
     /**
      * @brief Get the shape representing the fruit
@@ -44,7 +44,7 @@ public:
     sf::CircleShape getShape() && noexcept;
 
 private:
-    sf::Vector2i m_position {};
+    sf::Vector2f m_position {};
     sf::CircleShape m_shape {};
 };
